refactor(RangeSlider): Move styled-components into partials
