+++ packages/react-component-library/src/components/RangeSlider/partials/StyledIconRight.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+import { selectors } from '@royalnavy/design-tokens'
+
+const { color, spacing } = selectors
+
+export const StyledIconRight = styled.div`
+  svg {
+    color: ${color('neutral', '400')};
+    overflow: visible;
+    margin-left: ${spacing('2')};
+  }
+`
