+++ packages/react-component-library/src/components/RangeSlider/partials/StyledRail.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const StyledRail = styled.div`
+  position: absolute;
+  top: calc(50% + 20px);
+  left: 0;
+  display: inline-block;
+  width: 100%;
+  height: 40px;
+  transform: translateY(-100%);
+`
