+++ packages/react-component-library/src/components/RangeSlider/partials/StyledRangeSlider.tsx
@@ -1,0 +1,24 @@
+import styled, { css } from 'styled-components'
+
+interface StyledRangeSliderProps {
+  $isReversed?: boolean
+  $isDisabled?: boolean
+  $hasPercentage?: boolean
+}
+
+export const StyledRangeSlider = styled.div<StyledRangeSliderProps>`
+  display: inline-flex;
+  align-items: center;
+  width: 100%;
+
+  ${({ $isDisabled }) =>
+    $isDisabled &&
+    css`
+      opacity: 0.5;
+      cursor: not-allowed;
+
+      * {
+        cursor: not-allowed;
+      }
+    `}
+`
