+++ packages/react-component-library/src/components/RangeSlider/partials/StyledTicks.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+export const StyledTicks = styled.div`
+  div:first-of-type,
+  div:last-of-type {
+    div {
+      height: 16px;
+    }
+  }
+`
